Refactor : Move spotify elements
